:gem: accounts -> cadastro, update da propria conta ou de terceiros, delete de conta com limitação para nao deletar a mesma conta que esta logada

+++ api/controllers/login.go
@@ -56,7 +56,7 @@ func Login(c *gin.Context) {
 		msgreturn = "invalid credentials"
 	} else {
 
-		token, err := services.NewJWTService().GenerateToken(uint(p.Account))
+		token, err := services.NewJWTService().GenerateToken((p.Account))
 
 		if err != nil {
 			code = 500
